Adding Croatian language
Unfortunately I had to add the language option to the end since the
Prefs seems to be matched on index position in the drop down list
instead of the key. I'll look into that in another commit.

But if I would have added Croatian on the correct alphabetical position
all current installations that updated would get one language offset on
all languages after Croatian.